Add NO_TARGET option to build time copy helper.
By default cpp_cc_build_time_copy creates a target that is always built
that depends on the output file, ensuring that it is already up to date.
This potentially generates a large number of top-level always-built
targets, which can be rather slow when using legacy make/Makefiles. The
Ninja build system does not seem to have the same problem. This commit
adds an optional NO_TARGET flag, which sets up the rule to copy the file
but does not create a top-level always-built target. This means that the
caller can create a single such target depending on many files, which is
much more efficient when using make/Makefiles.