Fix infinite os.execv when unsuitable dependencies are specified in PYTHONPATH (#151)
Remove incompatible Python tools from environment

Whenever a Python tool is about to be installed with `pip`
in the virtualenv, get rid of previous installation of this tool
in the environment if any (sys.path and PYTHONPATH environment variable)
so that it doesn't interference with pip installation.
The change also detects the conflicts encountered by pip to ensure
that unsuitable dependencies of the package are also excluded from
the environment.

This change solves an issue where an unsuitable version of black was provided
via PYTHONPATH, and pip was not able to uninstall it because provided from
outside. The required version of black also needed a newer version of `pathspec`
provided in PYTHONPATH. Therefore, despite the `pip install proper-version-of-black`,
`pip` considered that the conditions were not met, and the loop was going on forever.